Update scripts/RELEASE_CHECKLIST.md
The list has had several changes - modifications relevant to the switch
between GitHub and Gitlab, and commit tagging information is now noted.

Importantly, the "Packaging Guide" was removed because it didn't add
anything to the document that wasn't already in the checklist, and the
"PyPI Credentials" section was removed because storing credentials in
plaintext is a bad idea so following it makes little sense.